requirements.txt
Django==1.10.6
psycopg2==2.7
django-geoposition==0.3.0
django-jsonview==1.0.0
requests==2.13.0
